2013_11_29 - 10:52pm update
John - when I input a question, I'm now able to output a formatted form
with the question in the textarea.  I can also append one question onto
an existing question in the text area (not shown).

Next: Setting up a function with a loop so each question gets is own
output form...